requirements.txt
pytest
selenium
ipython
allure-pytest
allure-python-commons
nose
